+++ db/migrate/20120531093420_create_sku_price_logs.rb
@@ -2,8 +2,8 @@ class CreateSkuPriceLogs < ActiveRecord::Migration
   def change
     create_table :sku_price_logs do |t|
       t.integer :sku_id,  :null => false,   :on_delete => :cascade
-      t.integer :before,  :null => false,   :limit => 5
-      t.integer :after,   :null => false,   :limit => 5
+      t.float   :before,  :null => false,   :limit => 5
+      t.float   :after,   :null => false,   :limit => 5
 
       t.user_tracking
       t.timestamps
